+++ www/Library/content_widgets/Title/Title.php
@@ -12,8 +12,8 @@ class Title extends Widget
 {
     public function work($v = array())
     {
-        $v['value'] = $this->_input['GET']['object']->getValue();
-        if ($parent = $this->_input['GET']['object']->parent()){
+        $v['value'] = $this->_input['REQUEST']['object']->getValue();
+        if ($parent = $this->_input['REQUEST']['object']->parent()){
             $v['parent_uri'] = $parent['uri'];
             if (mb_substr($v['parent_uri'],0,10)=='/Contents/'){
                 $v['parent_uri'] = mb_substr($v['parent_uri'],10);