+++ www/Library/layouts/Admin/Programs/Programs.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Класс
+ *
+ * @version 1.0
+ */
+namespace Library\layouts\Admin\Programs;
+
+use Library\views\AutoWidget\AutoWidget;
+
+class Programs extends AutoWidget
+{
+    protected function initInput($input)
+    {
+        return parent::initInput($input);
+    }
+
+    public function work($v = array())
+    {
+        return parent::work($v);
+    }
+}