+++ www/Library/views/PageNavigation/PageNavigation.php
@@ -14,7 +14,7 @@ class PageNavigation extends Widget
     public function getInputRule()
     {
         return Rule::arrays(array(
-            'GET' => Rule::arrays(array(
+            'REQUEST' => Rule::arrays(array(
                 'object' => Rule::entity()->required(), // объект, который отображается постранично
                 'page' => Rule::int()->default(1)->required(),
                 'page_count' => Rule::int()->default(1)->required()
@@ -24,10 +24,10 @@ class PageNavigation extends Widget
 
     public function work($v = array())
     {
-        $obj = $this->_input['GET']['object'];
+        $obj = $this->_input['REQUEST']['object'];
         $v['uri'] = substr($obj['uri'], 9);
-        $v['count'] = $this->_input['GET']['page_count'];
-        $v['current'] = min($v['count'], $this->_input['GET']['page']);
+        $v['count'] = $this->_input['REQUEST']['page_count'];
+        $v['current'] = min($v['count'], $this->_input['REQUEST']['page']);
         return parent::work($v);
     }
 }